trial-discovery-ai/backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "acquittify-peregrine-backend"
version = "0.1.0"
description = "Acquittify Peregrine backend"
requires-python = ">=3.12"
readme = "README.md"
dependencies = [
  "fastapi>=0.111.0",
  "uvicorn>=0.30.0",
  "passlib>=1.7.4",
  "pydantic-settings>=2.3.0",
  "sqlalchemy>=2.0.30",
  "alembic>=1.13.2",
  "psycopg[binary]>=3.1.18",
  "pgvector>=0.2.5",
  "dropbox>=11.36.2",
  "boto3>=1.34.131",
  "openpyxl>=3.1.5",
  "python-docx>=1.1.2",
  "pillow-heif>=0.18.0",
  "pymupdf>=1.24.7",
  "pytesseract>=0.3.10",
  "pillow>=10.4.0",
  "langdetect>=1.0.9",
  "mail-parser>=3.15.0",
  "requests>=2.32.0",
  "openai>=1.40.0",
  "pyotp>=2.9.0",
  "cryptography>=43.0.0",
  "jsonschema>=4.22.0",
  "celery>=5.3.6",
  "redis>=5.0.0",
  "python-multipart>=0.0.22",
  "python-docx>=1.2.0",
  "openpyxl>=3.1.5",
]

[project.optional-dependencies]
dev = [
  "httpx>=0.27.0",
  "pytest>=8.3.0",
]

[tool.uvicorn]
factory = false

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
include = ["app*"]
exclude = ["Casefiles*", "tests*", "scripts*"]
